This Feature allows the user to register to the database upon button click
register_user.jsp is the jsp file responsible for storing the data to the database
