+++ nbproject/private/private.properties
@@ -1,5 +1,5 @@
-deploy.ant.properties.file=/Users/danielflachica/Library/Application Support/NetBeans/8.2/tomcat80.properties
-j2ee.server.home=/Library/Tomcat/apache-tomcat-8.5.28
-j2ee.server.instance=tomcat80:home=/Library/Tomcat/apache-tomcat-8.5.28
-user.properties.file=/Users/danielflachica/Library/Application Support/NetBeans/8.2/build.properties
+deploy.ant.properties.file=C:\\Users\\Marl\\AppData\\Roaming\\NetBeans\\8.2\\tomcat70.properties
+j2ee.server.home=C:/xampp/tomcat
+j2ee.server.instance=tomcat70:home=C:\\xampp\\tomcat
+user.properties.file=C:\\Users\\Marl\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
