--- src/java/jdbc/Login.java
@@ -1,38 +1,0 @@
-package jdbc;
-
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/**
- *
- * @author Chelsey
- */
-@WebServlet(urlPatterns = {"/Login"})
-public class Login extends HttpServlet {
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException{
-        
-        String uname = request.getParameter("uname");
-        String password = request.getParameter("pword");
-    
-        if(uname.equals("chels") && password.equals("wassup")){
-                response.sendRedirect("index.htm");
-        }else{
-                response.sendRedirect("login.jsp");
-        }
-    }
-}
